move content analysis model back to models
