+++ lexos/receivers/contentanalysis_receiver.py
