+++ test/unit_test/test_content_analysis.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from lexos.processors.analyze.content_analysis_model import \
+from lexos.models.content_analysis_model import \
     ContentAnalysisModel
